backend/requirements.txt
Flask
Flask-Cors
mysql-connector-python
python-dotenv
bcrypt